Create default dotfiles
The initial commit of this repository features the creation of some
standard dotfiles I use on my system. This includes:

 - .bash_aliases
 - .bashrc
 - .gitconfig
 - vimrc

Excluding a few aliases and a custom terminal status line, vimrc is the
bulk of this commit (and repository). The executable 'copy' can be run
to automatically place the files in their directories.